chore(ui): Move helpers and selectors into vulnmanagement tests folder
