+++ ui/apps/platform/cypress/integration/vulnmanagement/VulnerabilityManagement.helpers.js
@@ -1,9 +1,9 @@
-import { selectors } from '../../constants/VulnManagementPage';
-import { hasFeatureFlag } from '../features';
+import { hasFeatureFlag } from '../../helpers/features';
+import { visitFromLeftNavExpandable } from '../../helpers/nav';
+import { getRouteMatcherMapForGraphQL, interactAndWaitForResponses } from '../../helpers/request';
+import { visit } from '../../helpers/visit';
 
-import { visitFromLeftNavExpandable } from '../nav';
-import { getRouteMatcherMapForGraphQL, interactAndWaitForResponses } from '../request';
-import { visit } from '../visit';
+import { selectors } from './VulnerabilityManagement.selectors';
 
 const opnamesForDashboard = [
     'cvesCount',
@@ -237,36 +237,27 @@ export function interactAndWaitForVulnerabilityManagementSecondaryEntities(
     interactAndWaitForResponses(interactionCallback, routeMatcherMap, staticResponseMap);
 }
 
-/*
- * resultsFromRegExp: /^(\d+) (\D+)$/.exec(linkText)
- * which assumes that linkText matches a more specific RegExp
- * for example, /^\d+ deployments?$/
- */
+// The one-based index includes checkbox, hidden, invisible cells.
+const getTableDataColumnSelector = (columnIndex) => `.rt-tbody .rt-td:nth-child(${columnIndex})`;
 
-// After accessibility-related changes to case of entity types,
-// getCountAndNoun functions might become obsolete, as follows:
+export function visitVulnerabilityManagementSecondaryEntitiesInSidePanel(
+    entitiesKey1,
+    entitiesKey2,
+    columnIndex // one-based index includes checkbox, hidden, invisible
+) {
+    // 1. Visit list page for primary entities.
+    visitVulnerabilityManagementEntities(entitiesKey1);
 
-// 1. Done for ROX-17001
-
-// 2. TODO because visible text is better than data-testid attribute.
-//    Replace selector which has data-testid attribute
-//    with contains method and RegExp for exact match:
-//    correct case entity type noun with optional plural suffix.
-
-export function getCountAndNounFromImageCVEsLinkResults([, count]) {
-    return {
-        panelHeaderText: `${count} Image ${count === '1' ? 'CVE' : 'CVES'}`,
-        relatedEntitiesCount: count,
-        relatedEntitiesNoun: count === '1' ? 'IMAGE CVE' : 'IMAGE CVES',
-    };
-}
-
-export function getCountAndNounFromNodeCVEsLinkResults([, count]) {
-    return {
-        panelHeaderText: `${count} Node ${count === '1' ? 'CVE' : 'CVES'}`,
-        relatedEntitiesCount: count,
-        relatedEntitiesNoun: count === '1' ? 'NODE CVE' : 'NODE CVES',
-    };
+    // 2. Click the first link for secondary entities to open side panel.
+    interactAndWaitForVulnerabilityManagementSecondaryEntities(
+        () => {
+            cy.get(getTableDataColumnSelector(columnIndex))
+                .contains('a', countNounRegExp[entitiesKey2])
+                .click();
+        },
+        entitiesKey1,
+        entitiesKey2
+    );
 }
 
 // TODO menuListItemRegExp needs only because sentence case instead of ordinary case.
@@ -293,14 +284,6 @@ export function verifyVulnerabilityManagementDashboardCVEs(entitiesKey, menuList
  * /main/vulnerability-management/namespace/id/deployments
  *
  * columnIndex is one-based but would be dataLabel in PatternFly.
- *
- * entitiesRegExp2 matches links in primary entities table to secondary entities.
- * For example, /^\d+ deployments?$/
- *
- * getCountAndNounFromLinkResults optioanl function provides the noun.
- * For example,
- * Noun is not in link text: /^\d+ CVEs?$/ or /^\d+ Fixable$/
- * Noun differs from link text: /^\d+ failing deployments?$/
  */
 export function verifySecondaryEntities(
     entitiesKey1,
@@ -316,6 +299,10 @@ export function verifySecondaryEntities(
 
 /*
  * Verify panelHeader text, and then visit related entities pages,
+ *
+ * entitiesRegExp2 matches links in primary entities table to secondary entities.
+ * Noun is not in link text: /^\d+ CVEs?$/ or /^\d+ Fixable$/
+ * Noun differs from link text: /^\d+ failing deployments?$/
  */
 function verifyTableLink(
     entitiesKey1,
@@ -324,7 +311,7 @@ function verifyTableLink(
     entitiesRegExp2
 ) {
     // Find the first link for secondary entities.
-    cy.get(selectors.getTableDataColumnSelector(columnIndex))
+    cy.get(getTableDataColumnSelector(columnIndex))
         .contains('a', entitiesRegExp2 ?? countNounRegExp[entitiesKey2])
         .then(($a) => {
             // 2. Visit secondary entities side panel.
@@ -339,7 +326,7 @@ function verifyTableLink(
 
             // 3. Visit primary entity side panel.
             interactAndWaitForResponses(() => {
-                cy.get(selectors.parentEntityInfoHeader).click();
+                cy.get('[data-testid="breadcrumb-link-text"] > a').click();
             }, getRouteMatcherMapForGraphQL([opnameForEntity[entitiesKey1]]));
 
             // Tilde because link might be under either Contains or Matches.
@@ -397,27 +384,27 @@ export function verifyConditionalCVEs(
 
         if (hasFixableCVEs) {
             // If at least one of entitiesKey1 has fixable CVEs, then CVEs link exists.
-            cy.get(selectors.getTableDataColumnSelector(columnIndex))
+            cy.get(getTableDataColumnSelector(columnIndex))
                 .contains('a', allCVEsRegExp)
                 .should('exist');
 
             verifyTableLink(entitiesKey1, entitiesKey2, columnIndex, fixableCVEsRegExp);
         } else if (hasCVEs) {
             // Fixable link does not exist in any row of entityKeys1 list.
-            cy.get(selectors.getTableDataColumnSelector(columnIndex))
+            cy.get(getTableDataColumnSelector(columnIndex))
                 .contains('a', fixableCVEsRegExp)
                 .should('not.exist');
 
             verifyTableLink(entitiesKey1, entitiesKey2, columnIndex, allCVEsRegExp);
         } else {
             // Neither link exists in any row of entitiesKey1 list.
-            cy.get(selectors.getTableDataColumnSelector(columnIndex))
+            cy.get(getTableDataColumnSelector(columnIndex))
                 .contains('a', fixableCVEsRegExp)
                 .should('not.exist');
-            cy.get(selectors.getTableDataColumnSelector(columnIndex))
+            cy.get(getTableDataColumnSelector(columnIndex))
                 .contains('a', allCVEsRegExp)
                 .should('not.exist');
-            cy.get(`${selectors.getTableDataColumnSelector(columnIndex)}:contains("No CVEs")`);
+            cy.get(`${getTableDataColumnSelector(columnIndex)}:contains("No CVEs")`);
         }
     });
 }
