+++ ui/apps/platform/cypress/integration/vulnmanagement/VulnerabilityManagement.selectors.js
@@ -1,0 +1,36 @@
+export const listSelectors = {
+    tableRows: '.rt-tr',
+    tableBodyRowGroups: '.rt-tbody .rt-tr-group',
+    tableBodyRows: '.rt-tbody .rt-tr',
+    tableRowCheckbox: '[data-testid="checkbox-table-row-selector"]',
+    cveDescription: '[data-testid="cve-description"]',
+    metadataDescription: '[data-testid="metadata-description"]',
+    statusChips: '[data-testid="label-chip"]',
+    failingDeploymentCountLink: '[data-testid="failingDeploymentsCountLink"]',
+    cveAddToPolicyButton: '[data-testid="panel-button-add-cves-to-policy"]',
+    cveAddToPolicyShortForm: {
+        // TODO: fix the following selector for react-select, that evil component
+        select: '[data-testid="policy-short-form"] select',
+        selectValue: '[data-testid="policy-short-form"] .react-select__multi-value__label',
+    },
+};
+
+export const dashboardSelectors = {
+    getWidget: (title) => {
+        return `[data-testid="widget"]:contains('${title}')`;
+    },
+};
+
+const sidePanelSelectors = {
+    sidePanel: '[data-testid="side-panel"]',
+    backButton: '[data-testid="sidepanelBackButton"]',
+    entityIcon: '[data-testid="entity-icon"]',
+    sidePanelExternalLinkButton: '[aria-label="External link"]',
+    scanDataMessage: '[data-testid="message"].error-message:contains("CVE Data May Be Inaccurate")',
+};
+
+export const selectors = {
+    ...dashboardSelectors,
+    ...listSelectors,
+    ...sidePanelSelectors,
+};
