+++ ui/apps/platform/cypress/integration/vulnmanagement/clusters.test.js
@@ -6,22 +6,12 @@ import {
     callbackForPairOfDescendingNumberValuesFromElements,
 } from '../../helpers/sort';
 import {
-    getCountAndNounFromImageCVEsLinkResults,
-    getCountAndNounFromNodeCVEsLinkResults,
     hasTableColumnHeadings,
     interactAndWaitForVulnerabilityManagementEntities,
     verifyConditionalCVEs,
     verifySecondaryEntities,
     visitVulnerabilityManagementEntities,
-} from '../../helpers/vulnmanagement/entities';
-
-function getCountAndNounFromClusterCVEsLinkResults([, count]) {
-    return {
-        panelHeaderText: `${count} Platform ${count === '1' ? 'CVE' : 'CVES'}`,
-        relatedEntitiesCount: count,
-        relatedEntitiesNoun: count === '1' ? 'CLUSTER CVE' : 'CLUSTER CVES',
-    };
-}
+} from './VulnerabilityManagement.helpers';
 
 const entitiesKey = 'clusters';
 
@@ -85,13 +75,7 @@ describe('Vulnerability Management Clusters', () => {
     // The one-based index includes checkbox, hidden, invisible.
 
     it('should display either links for image CVEs or text for No CVEs', () => {
-        verifyConditionalCVEs(
-            entitiesKey,
-            'image-cves',
-            3,
-            'imageVulnerabilityCounter',
-            getCountAndNounFromImageCVEsLinkResults
-        );
+        verifyConditionalCVEs(entitiesKey, 'image-cves', 3, 'imageVulnerabilityCounter');
     });
 
     it('should display either links for node CVEs or text for No CVEs', function () {
@@ -99,13 +83,7 @@ describe('Vulnerability Management Clusters', () => {
             this.skip(); // TODO verify and remove
         }
 
-        verifyConditionalCVEs(
-            entitiesKey,
-            'node-cves',
-            4,
-            'nodeVulnerabilityCounter',
-            getCountAndNounFromNodeCVEsLinkResults
-        );
+        verifyConditionalCVEs(entitiesKey, 'node-cves', 4, 'nodeVulnerabilityCounter');
     });
 
     it('should display either links for cluster CVEs or text for No CVEs', function () {
@@ -113,13 +91,7 @@ describe('Vulnerability Management Clusters', () => {
             this.skip(); // TODO verify and remove
         }
 
-        verifyConditionalCVEs(
-            entitiesKey,
-            'cluster-cves',
-            5,
-            'clusterVulnerabilityCounter',
-            getCountAndNounFromClusterCVEsLinkResults
-        );
+        verifyConditionalCVEs(entitiesKey, 'cluster-cves', 5, 'clusterVulnerabilityCounter');
     });
 
     it('should display links for namespaces', () => {
