+++ ui/apps/platform/cypress/integration/vulnmanagement/nodeComponents.test.js
@@ -6,13 +6,12 @@ import {
     callbackForPairOfDescendingNumberValuesFromElements,
 } from '../../helpers/sort';
 import {
-    getCountAndNounFromNodeCVEsLinkResults,
     hasTableColumnHeadings,
     interactAndWaitForVulnerabilityManagementEntities,
     verifyConditionalCVEs,
     verifySecondaryEntities,
     visitVulnerabilityManagementEntities,
-} from '../../helpers/vulnmanagement/entities';
+} from './VulnerabilityManagement.helpers';
 
 const entitiesKey = 'node-components';
 
@@ -123,13 +122,7 @@ describe('Vulnerability Management Node Components', () => {
             this.skip(); // TODO verify and remove
         }
 
-        verifyConditionalCVEs(
-            entitiesKey,
-            'node-cves',
-            4,
-            'vulnCounter',
-            getCountAndNounFromNodeCVEsLinkResults
-        );
+        verifyConditionalCVEs(entitiesKey, 'node-cves', 4, 'vulnCounter');
     });
 
     it('should display links for nodes', () => {
